pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "mof"
authors = [{name = "Abbas Srour", email = "asrour@ciatek.net"}]
readme = "README.md"
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "fastapi >=0.70.1",
    "gunicorn >=20.1.0",
    "uvicorn >=0.16.0",
    "cryptography >=3.4.7",
    "authlib >=0.15.4",
    "PyJWT >=2.1.0",
    "passlib >=1.7.4",
    "email-validator >=1.1.3",
    "python-multipart >=0.0.5",
    "bcrypt >=3.2.0",
    "python-dateutil >=2.8.2",
    "python-dotenv >=0.20.0",
    "beanie >= 1.21.0b1",
    "fastapi-pagination >=0.12.0",
    "pydantic_settings >= 2.0.2",
    "pydantic-extra-types >= 2.0.0",
    "phonenumbers >= 8.13.18",
    "python-jose>= 3.3.0"
]

[project.optional-dependencies]
test = [
    "pytest >=7.4.0",
    "mypy >=0.910",
    "coverage >=6.1.1",
    "pytest-cov >=3.0.0",
]

dev = [
    "flake8 >=3.9.2",
    "black >=21.10b0",
    "isort >=5.9.3",
    "autoflake >=1.4",
    "flake8-docstrings >=1.6.0",
    "pre-commit >=2.17.0",
]

[project.urls]
Home = "https://github.com/ciatek/MOF"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"

